+++ capa/features/extractors/ghidra/helpers.py
@@ -1,16 +1,13 @@
-# Copyright (C) 2020 Mandiant, Inc. All Rights Reserved.
+# Copyright (C) 2023 Mandiant, Inc. All Rights Reserved.
 # Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at: [package root]/LICENSE.txt
 # Unless required by applicable law or agreed to in writing, software distributed under the License
 #  is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and limitations under the License.
-from typing import Any, Dict, Tuple, Iterator, Optional
+from typing import Iterator
 
 import ghidra
-from ghidra.program.model.symbol import SymbolType
-
-currentProgram: ghidra.program.database.ProgramDB
 
 
 def fix_byte(b: int) -> bytes:
@@ -30,9 +27,8 @@ def find_byte_sequence(seq: bytes) -> Iterator[int]:
     """
     seqstr = "".join([f"\\x{b:02x}" for b in seq])
     # .add(1) to avoid false positives on regular PE files
-    ea = findBytes(currentProgram.getMinAddress().add(1), seqstr, 1, 1)  # type: ignore [name-defined]
-    for e in ea:
-        yield e
+    eas = findBytes(currentProgram.getMinAddress().add(1), seqstr, 1, 1)  # type: ignore [name-defined] # noqa: F821
+    yield from eas
 
 
 def get_bytes(addr: ghidra.program.model.address.Address, length: int) -> bytes:
@@ -45,7 +41,7 @@ def get_bytes(addr: ghidra.program.model.address.Address, length: int) -> bytes:
 
     bytez = b""
     try:
-        signed_ints = getBytes(addr, length)  # type: ignore [name-defined]
+        signed_ints = getBytes(addr, length)  # type: ignore [name-defined] # noqa: F821
         for b in signed_ints:
             bytez = bytez + fix_byte(b)
         return bytez
@@ -62,7 +58,7 @@ def get_block_bytes(block: ghidra.program.model.mem.MemoryBlock) -> bytes:
 
     bytez = b""
     try:
-        signed_ints = getBytes(block.getStart(), block.getEnd().getOffset() - block.getStart().getOffset())  # type: ignore [name-defined]
+        signed_ints = getBytes(block.getStart(), block.getEnd().getOffset() - block.getStart().getOffset())  # type: ignore [name-defined] # noqa: F821
         for b in signed_ints:
             bytez = bytez + fix_byte(b)
         return bytez
@@ -73,5 +69,4 @@ def get_block_bytes(block: ghidra.program.model.mem.MemoryBlock) -> bytes:
 def get_function_symbols() -> Iterator[ghidra.program.database.function.FunctionDB]:
     """yield all non-external function symbols"""
 
-    for f in currentProgram.getFunctionManager().getFunctionsNoStubs(True):
-        yield f
+    yield from currentProgram.getFunctionManager().getFunctionsNoStubs(True)  # type: ignore [name-defined] # noqa: F821