+++ capa/helpers.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2020 Mandiant, Inc. All Rights Reserved.
+# Copyright (C) 2023 Mandiant, Inc. All Rights Reserved.
 # Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at: [package root]/LICENSE.txt
@@ -41,6 +41,10 @@ def get_file_taste(sample_path: Path) -> bytes:
 
 def is_runtime_ida():
     return importlib.util.find_spec("idc") is not None
+
+
+def is_runtime_ghidra():
+    return importlib.util.find_spec("ghidra.program.flatapi") is not None
 
 
 def assert_never(value) -> NoReturn: